+++ ru.taximaxim.codekeeper.core/src/main/java/ru/taximaxim/codekeeper/core/Utils.java
@@ -25,8 +25,15 @@ import java.nio.file.Paths;
 import java.util.List;
 import java.util.function.UnaryOperator;
 
+import javax.xml.XMLConstants;
 import javax.xml.parsers.DocumentBuilderFactory;
 import javax.xml.parsers.ParserConfigurationException;
+import javax.xml.transform.OutputKeys;
+import javax.xml.transform.Transformer;
+import javax.xml.transform.TransformerException;
+import javax.xml.transform.TransformerFactory;
+import javax.xml.transform.dom.DOMSource;
+import javax.xml.transform.stream.StreamResult;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -77,12 +84,27 @@ public final class Utils {
         factory.setFeature("http://xml.org/sax/features/external-parameter-entities", false); //$NON-NLS-1$
 
         // Prohibit the use of all protocols by external entities (JAXP 1.5+)
-        factory.setAttribute(javax.xml.XMLConstants.ACCESS_EXTERNAL_DTD, ""); //$NON-NLS-1$
-        factory.setAttribute(javax.xml.XMLConstants.ACCESS_EXTERNAL_SCHEMA, ""); //$NON-NLS-1$
+        factory.setAttribute(XMLConstants.ACCESS_EXTERNAL_DTD, ""); //$NON-NLS-1$
+        factory.setAttribute(XMLConstants.ACCESS_EXTERNAL_SCHEMA, ""); //$NON-NLS-1$
 
         Document doc = factory.newDocumentBuilder().parse(new InputSource(reader));
         doc.normalize();
         return doc;
+    }
+
+    public static void writeXml(Document xml, boolean encrypt, StreamResult stream) throws TransformerException {
+        TransformerFactory factory = TransformerFactory.newInstance();
+        factory.setAttribute(XMLConstants.ACCESS_EXTERNAL_DTD, ""); //$NON-NLS-1$
+        factory.setAttribute(XMLConstants.ACCESS_EXTERNAL_STYLESHEET, ""); //$NON-NLS-1$
+
+        if (encrypt) {
+            factory.setFeature(XMLConstants.FEATURE_SECURE_PROCESSING, true);
+        }
+
+        Transformer tf = factory.newTransformer();
+        tf.setOutputProperty(OutputKeys.INDENT, "yes"); //$NON-NLS-1$
+        tf.setOutputProperty("{http://xml.apache.org/xslt}indent-amount", "4"); //$NON-NLS-1$ //$NON-NLS-2$
+        tf.transform(new DOMSource(xml), stream);
     }
 
     public static boolean isSystemSchema(String schema, DatabaseType dbType) {