+++ ru.taximaxim.codekeeper.core/src/main/java/ru/taximaxim/codekeeper/core/xmlstore/XmlStore.java
@@ -26,14 +26,9 @@ import java.text.MessageFormat;
 import java.util.ArrayList;
 import java.util.List;
 
-import javax.xml.XMLConstants;
 import javax.xml.parsers.DocumentBuilderFactory;
 import javax.xml.parsers.ParserConfigurationException;
-import javax.xml.transform.OutputKeys;
-import javax.xml.transform.Transformer;
 import javax.xml.transform.TransformerException;
-import javax.xml.transform.TransformerFactory;
-import javax.xml.transform.dom.DOMSource;
 import javax.xml.transform.stream.StreamResult;
 
 import org.w3c.dom.Document;
@@ -110,12 +105,16 @@ public abstract class XmlStore<T> {
         try {
             Path path = getXmlFile();
             Files.createDirectories(path.getParent());
-            try (Writer xmlWriter = Files.newBufferedWriter(path, StandardCharsets.UTF_8)) {
-                serializeXml(xml, true, xmlWriter);
-            }
-        } catch (IOException | TransformerException ex) {
+            writeDocument(xml, path);
+        } catch (Exception ex) {
             throw new IOException(MessageFormat.format(
                     Messages.XmlStore_write_error, ex.getLocalizedMessage()), ex);
+        }
+    }
+
+    protected void writeDocument(Document xml, Path path) throws IOException, TransformerException {
+        try (Writer xmlWriter = Files.newBufferedWriter(path, StandardCharsets.UTF_8)) {
+            Utils.writeXml(xml, false, new StreamResult(xmlWriter));
         }
     }
 
@@ -131,34 +130,24 @@ public abstract class XmlStore<T> {
         if (useCached && cachedDocument != null) {
             return cachedDocument;
         }
+        Document xml = readXml();
+        if (!xml.getDocumentElement().getNodeName().equals(rootTag)) {
+            throw new IOException(Messages.XmlStore_root_error);
+        }
+        cachedDocument = xml;
+        return xml;
+    }
+
+    protected Document readXml() throws IOException {
         try (Reader reader = Files.newBufferedReader(getXmlFile(), StandardCharsets.UTF_8)) {
-            Document xml = Utils.readXml(reader);
-
-            if (!xml.getDocumentElement().getNodeName().equals(rootTag)) {
-                throw new IOException(Messages.XmlStore_root_error);
-            }
-
-            cachedDocument = xml;
-            return xml;
+            return Utils.readXml(reader);
         } catch (NoSuchFileException ex) {
             throw ex;
         } catch (IOException | SAXException | ParserConfigurationException ex) {
             throw new IOException(MessageFormat.format(
                     Messages.XmlStore_read_error, ex.getLocalizedMessage()), ex);
+        } catch (Exception e) {
+            throw new IOException(MessageFormat.format(Messages.XmlStore_read_error, e.getLocalizedMessage()), e);
         }
-    }
-
-    private void serializeXml(Document xml, boolean formatting,
-            Writer writer) throws TransformerException {
-        TransformerFactory factory = TransformerFactory.newInstance();
-        factory.setAttribute(XMLConstants.ACCESS_EXTERNAL_DTD, "");
-        factory.setAttribute(XMLConstants.ACCESS_EXTERNAL_STYLESHEET, "");
-
-        Transformer tf = factory.newTransformer();
-        if (formatting) {
-            tf.setOutputProperty(OutputKeys.INDENT, "yes"); //$NON-NLS-1$
-            tf.setOutputProperty("{http://xml.apache.org/xslt}indent-amount", "4"); //$NON-NLS-1$ //$NON-NLS-2$
-        }
-        tf.transform(new DOMSource(xml), new StreamResult(writer));
     }
 }