+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
-- Disabled ability to save passwords outside Eclipse secure storage.
+- Added encryption of the dbstore.xml file.
 - Added checks when deserializing objects.
 - Improved work with created temporary files.
 - Increased version of JDBC driver for PostgreSQL.